Need more space for the year, changed column to smallint

+++ Tables/dbo.SiteStats.sql
@@ -3,7 +3,7 @@ CREATE TABLE [dbo].[SiteStats]
 [StatID] [int] NOT NULL IDENTITY(1, 1),
 [StateDate] [date] NULL CONSTRAINT [DF__SiteStats__State__4589517F] DEFAULT (sysdatetime()),
 [StatMonth] [tinyint] NULL,
-[StatYear] [tinyint] NULL,
+[StatYear] [smallint] NULL,
 [PageVisits] [int] NULL,
 [TimeOnSite] [time] NULL,
 [Engagement] AS ([PageVisits]*datediff(second,CONVERT([time],'00:00:00',(0)),[TimeOnSite]))
